Update README.md
Art is about the cycle of recreation. All art is built and inspired by other art. 🎨
In the case of Markov chains we are taking something apart and rebuilding it.✂️
Art is also what ever you define as art. To me this is fitting. 🖼️
It is the beauty of logic. ✨